refactor: Update user table columns in migration
This commit updates the user table columns in the migration file. It removes the "roles" column and the "first_name" and "last_name" columns, and adds a new "full_name" column. It also changes the "role" column to use a new enum type "users_role_enum" with default value "USER".

Refactor the user table columns in the migration file to reflect the changes in the database schema.